requirements.txt
Django==1.6.1
MySQL-python==1.2.5
PyYAML==3.10
Werkzeug==0.9.4
dj-static==0.0.5
django-allauth==0.15.0
django-debug-toolbar==1.0.1
django-extensions==1.3.3
django-newrelic-extensions==0.0.1
flup==1.0.2
gunicorn==18.0
ipython==1.1.0
newrelic==2.12.0.10
oauthlib==0.6.1
pystache==0.5.3
python-openid==2.2.5
requests==2.2.1
requests-oauthlib==0.4.0
six==1.5.2
sqlparse==0.1.10
static==1.0.2
uWSGI==2.0.1
wsgiref==0.1.2
